@@ -3,14 +3,15 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "node-sass": "^4.14.1",
+    "node-sass": "^7.0.1",
     "npm-run-all": "^4.1.3",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
-    "react-medium-image-zoom": "^4.3.1",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "^3.4.1",
-    "styled-components": "^4.4.1"
+    "nth-check": "^2.1.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-medium-image-zoom": "^4.4.3",
+    "react-router-dom": "^6.3.0",
+    "react-scripts": "^5.0.1",
+    "styled-components": "^5.3.5"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -30,7 +31,7 @@
     "not op_mini all"
   ],
   "devDependencies": {
-    "babel-plugin-styled-components": "^1.10.7"
+    "babel-plugin-styled-components": "^2.0.7"
   },
   "babelMacros": {
     "styledComponents": {}
